Move ServiceMode to iris.cluster.service_mode to break circular import
The import chain config.py → k8s/__init__.py → k8s_service_impl.py →
platform.service_mode triggered platform/__init__.py which imported
coreweave.py back into config.py. Moving service_mode.py up one level
to iris.cluster breaks the cycle.
