+++ lib/iris/src/iris/cluster/k8s/__init__.py
@@ -5,7 +5,7 @@
 
 from iris.cluster.k8s.k8s_service import K8sService
 from iris.cluster.k8s.k8s_service_impl import K8sServiceImpl
-from iris.cluster.platform.service_mode import ServiceMode
+from iris.cluster.service_mode import ServiceMode
 from iris.cluster.k8s.kubectl import Kubectl, KubectlError
 
 __all__ = ["K8sService", "K8sServiceImpl", "Kubectl", "KubectlError", "ServiceMode"]
