+++ lib/iris/src/iris/cluster/local_cluster.py
@@ -41,7 +41,7 @@ from iris.cluster.controller.scaling_group import (
 from iris.cluster.platform.base import find_free_port
 from iris.cluster.platform.gcp import GcpPlatform
 from iris.cluster.platform.gcp_service_impl import GcpServiceImpl
-from iris.cluster.platform.service_mode import ServiceMode
+from iris.cluster.service_mode import ServiceMode
 from iris.cluster.worker.port_allocator import PortAllocator
 from iris.managed_thread import ThreadContainer
 from iris.rpc import config_pb2
